Move app sources to other directory
