+++ Source/App/Inc/app.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void app(void);
