+++ Source/App/Inc/detector.h
