+++ Source/App/Inc/nv.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "defs.h"
+#include <stdint.h>
+
+void nvInit(void);
+UserSource nvGetState(void);
+void nvSetState(UserSource value);
