+++ Source/App/Src/detector.c
