--- Source/Core/Inc/nv.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "defs.h"
-#include <stdint.h>
-
-void nvInit(void);
-UserSource nvGetState(void);
-void nvSetState(UserSource value);
