+++ switch-firmware.code-workspace
@@ -26,7 +26,8 @@
 			"xtree": "c",
 			"xutility": "c",
 			"format": "c",
-			"router.h": "c"
+			"router.h": "c",
+			"remote.h": "c"
 		},
 		"cortex-debug.liveWatchRefreshRate": 2000,
 		"cortex-debug.dbgServerLogfile": "${workspaceRoot}/jlink.log",
